+++ database/migrations/2025_09_04_203020_create_expenses_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('expenses', function (Blueprint $table) {
             $table->id();
             $table->string('name')->nullable();
-            $table->float('amount')->nullable();
+            $table->decimal('amount',10,2)->nullable();
             $table->bigInteger('status')->default(0);
             $table->timestamps();
         });
